+++ lib/default.nix
@@ -1,6 +1,6 @@
-{ system, version, files, secrets, lib, inputs, pkgs, ... }:
+{ system, version, files, lib, inputs, pkgs, ... }:
 rec
 {
-  device = import ./device.nix { inherit system version files secrets lib user inputs pkgs; };
-  user = import ./user.nix { inherit system version files secrets lib inputs pkgs; };
+  device = import ./device.nix { inherit system version files lib user inputs pkgs; };
+  user = import ./user.nix { inherit system version files lib inputs pkgs; };
 }