+++ modules/hardware/default.nix
@@ -10,9 +10,20 @@ rec
     ./virtualisation.nix
   ];
 
-  options.hardware.cores = lib.mkOption
+  options.hardware =
   {
-    description = "Number of CPU Cores";
-    type = lib.types.int;
+    cores = lib.mkOption
+    {
+      description = "Number of CPU Cores";
+      type = lib.types.int;
+      default = 4;
+    };
+
+    support = lib.mkOption
+    {
+      description = "List of Additional Supported Hardware";
+      type = lib.types.listOf lib.types.str;
+      default = [ ];
+    };
   };
 }